Fixes issue-7892 with consolidated test (#7902)
